requirements.txt
Flask
requests
beautifulsoup4
pdfminer.six
